huffjoint: faster 4-wide joint table building
Most of the codewords are actually too large to fit with the LUT limit,
meaning wasteful loops are ran. By restricting the maximum symbol to 8,
the sign extending test can be skipped. Also, it is unlikely for
such small symbols to have unordered codelengths, thus breaking early
avoid useless loops.

The rationale for these was confirmed experimentally by checking the
number of elements in the tables.

This unfortunately can't be applied to 2-wide joint tables. This was
verified experimentally, and also most often, the most frequent VLC
pairs actually already fit.

Timings (kcycles) on 1024 runs (magicyuv): 48k -> 3k.