huffjoint: 4-element-wide joint tables
The implementation reuses the 2-wide joint tables. A later test on
reducing the number of elements tested for inclusion in these tables
showed an important slowdown, so they are likely still useful in
addition to the new 4-element-wide ones.

No implementation for >8-bits. Validated on magicyuv and huffyuvdec.

Some smarter granularity (3-wide?) could achieve better results but at
the cost of a likely even more complex implementation.