magicyuv: summarize 2-joint and 4-joint tables
Entries in the 4-joint table take priority over the ones in the 2-joint
one, thus they can be merged. At the same time, the data they old can be
shuffled so as to optimize their usage.

The LUT for 4 entries no longer needs to be kept around, so remove it
as well from the context.

Around 3-5% total speedup.