huffjoint: up to 6 symbols
There is that much room for 8 bits. However, table building was too slow
to make this worthwhile, i.e. 2-bit symbols must exist, indicating a
rather smooth content. It was however measured not to be detrimental
for content with artificially added noise.

Similarly, that allows decoding up to 3 symbols for more than 8 bits content.

Because the default macro can read up to 6 symbols for 8 bits (and 3 for more),
reducing the upper bound in loops is required.

Around 4-6% on content without noise, depending on codec (and <1% otherwise).