huffjoint: use a separate table for (more) symbols
This was benchmarked as faster for all codecs and bitdepths. In particular,
for >8 bits, this allows using up to 8 symbols (instead) of 3 previously,
albeit at the cost of a large data copy (16 bytes). The use of an 8 byte
copy for such content (and thuse 4 symbols at most) was found to be slower.

The cache usage actually decreases because, previously, a fixed 2^(bits+4)
bytes were used. Now, the amount is variable, but at the very least, for
entries, 2^(bits+2), plus LUT. The later was never found to be above 10KB.